requirements.txt
flipperzero==1.3.0
pyserial==3.5
protobuf==5.28.0
